This is weird! The iteration number was wrong.
I don’t know why. But this fixes the result-list.
The correct behavior is: We see iteration 148 in the URL and list and
we get the data that pivotal shows for this iteration 148.
I have no idea why we would need to subtract „2“ like this.

Lets move to the new API v5 soon … (see
https://github.com/JordanHatch/card-o-matic)